Do not raise state transition errors on PayPal transactions
Calling next! instead of next was causing StateMachine::InvalidTransition
errors to show when the transition failed. This caused 500 (internal
server) errors, which is not useful to the user. Instead the reason for
the failure such as decline or invalid address should have been
captured and shown to the user.

Example:
https://rollbar.com/fcp-euro/cloudspree/items/28595/